authentication and authorization

+++ src/modules/events/events.controller.ts
@@ -14,22 +14,23 @@ import {
   Inject,
   Query,
   UsePipes,
+  UseGuards,
+  ForbiddenException,
 } from '@nestjs/common';
 import { CreateEventDto } from './dto/create-event.dto';
 import { UpdateEventDto } from './dto/update-event.dto.';
-import { Event } from './event.entity';
-import { Repository } from 'typeorm';
-import { InjectRepository } from '@nestjs/typeorm';
 import { EventService } from './events.service';
 import { ListEvents } from './list.events';
+import { currentUser } from '../auth/current-user.decorator';
+import { AuthGuard } from '@nestjs/passport';
+import { User } from '../auth/user.entity';
+import { AuthGuardJwt } from '../auth/auth-guard.jwt';
 
 @Controller('/events')
 export class EventsController {
   private readonly logger = new Logger(EventsController.name);
 
   constructor(
-    @InjectRepository(Event)
-    private readonly repository: Repository<Event>,
     @Inject(EventService)
     private readonly eventService: EventService,
   ) {}
@@ -58,34 +59,64 @@ export class EventsController {
   }
 
   @Post()
-  create(@Body(ValidationPipe) input: CreateEventDto) {
-    return this.repository.save({
-      ...input,
-      when: new Date(input.when),
-    });
+  // @UseGuards(AuthGuard('jwt'))
+  @UseGuards(AuthGuardJwt)
+  create(
+    @currentUser() user: User,
+    @Body(ValidationPipe) input: CreateEventDto,
+  ) {
+    return this.eventService.createEvent(input, user);
   }
 
   @Patch(':id')
+  @UseGuards(AuthGuardJwt)
   async update(
     @Param('id', ParseIntPipe) id: number,
     @Body() input: UpdateEventDto,
+    @currentUser() user: User,
   ) {
-    const event = await this.repository.findOneBy({ id: id });
+    const event = await this.eventService.getEvent(id);
+    if (!event) {
+      throw new NotFoundException();
+    }
 
-    return this.repository.save({
-      ...event,
-      ...input,
-      when: input.when ? new Date(input.when) : event.when,
-    });
+    if (event.organizer_id !== user.id) {
+      throw new ForbiddenException(
+        null,
+        `you are not authorized to update this event`,
+      );
+    }
+
+    if (event.organizer_id !== user.id) {
+      throw new ForbiddenException(
+        null,
+        `you are not authorized to change this event`,
+      );
+    }
+
+    return this.eventService.updateEvent(event, input);
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuardJwt)
   @HttpCode(204)
-  async remove(@Param('id', ParseIntPipe) id: number) {
-    const result = await this.eventService.deleteEvent(id);
-    if (result.affected !== 1) {
+  async remove(
+    @Param('id', ParseIntPipe) id: number,
+    @currentUser() user: User,
+  ) {
+    const event = await this.eventService.getEvent(id);
+
+    if (!event) {
       throw new NotFoundException();
     }
-    return result;
+
+    if (event.organizer_id !== user.id) {
+      throw new ForbiddenException(
+        null,
+        `you are not authorized to delete this event`,
+      );
+    }
+
+    return this.eventService.deleteEvent(id);
   }
 }
